requirements.txt
django
psycopg2-binary
PyJWT
django-cors-headers